EoD
Have added NPCs to eventsList, but also noticed that numbers are not ascending correctly in either eventsList nor storyTeller. NPCs no showing where searchText is a comma seperated list, next steps now to merge Events and NPCs so that they have the same list and when you click on an NPC in events it jumps to npcForm and vice versa.