Instructions EoD
Have begun implementing a new method called instructions. These are tags added to Tags for items, etc that hold instructions to deploy to inheritors of that Tag. So for example, a Tag might have a random weapon tag so all inheritors recieve a random weapon.

This required changes to how tags are made, saved and loaded to make special rules for if the tag holds an instruction or not.

As of EoD, the tags can be created to the Tag if they are an item, and quantity and quality can be saved with them.

The code has been written in form.js to follow the instruction, I just have to decide where to put it. Do I want these items to be permanent or refresh every time character is saved?

I think I will have logic to permanently deploy instructions when the Tag us saved, but this will create problems with ever-increasing inventories.